Module fe_yulgen::names

Modules

abi

Functions

adjust_numeric_size

Generate a function name to adjust the size of the integer

checked_add

Generate a function name to perform checked addition

checked_div

Generate a function name to perform checked division

checked_exp

Generate a function name to perform checked exponentiation

checked_mod

Generate a function name to perform checked modulo

checked_mul

Generate a function name to perform checked multiplication

checked_neg

Generate a function name to perform checked negation

checked_sub

Generate a function name to perform checked subtraction

revert

Generates a revert function name for a given name and types

var_name

Generate a safe variable name for a user defined function