Module fe_yulgen::operations::abi

Enums

EncodingSize

The size of an encoding known at compile-time.

Functions

decode_component

Decode a single component.

decode_data

Decode a segment of memory and return each decoded component as separate values.

encode

Returns an expression that encodes the given values and returns a pointer to the encoding.

encoding_head_size

Returns an expression that gives the size of the encoding’s head.

encoding_known_size

Returns the known-at-compile-time encoding size.

encoding_size

Returns an expression that gives size of the encoded values.

unpack

Unpack each value into a newly allocated segment of memory.