Module fe_yulgen::runtime::functions::revert

Functions

all

Return all revert functions used by default.

error_revert

Generate a YUL function to revert with the Error signature and the given string.

error_revert_numeric

Generate a YUL function to revert with an error code.

panic_revert

Generate a YUL function to revert with a panic code.

revert

Generate a YUL function to revert with any signature name and type. Note: The parentheses on a tuple are removed in the selector preimage.