Module fe_yulgen::names

Modules

abi

Functions

adjust_numeric_size
Generate a function name to adjust the size of the integer

checked_add
Generate a function name to perform checked addition

checked_div
Generate a function name to perform checked division

checked_exp
Generate a function name to perform checked exponentiation

checked_mod
Generate a function name to perform checked modulo

checked_mul
Generate a function name to perform checked multiplication

checked_sub
Generate a function name to perform checked subtraction

contract_call
Generates an external call function name for a given type and location.

func_name
Generate a safe function name for a user defined function

revert
Generates a revert function name for a given name and types

struct_function_name
Generates a function name for to interact with a certain struct type

struct_getter_call
Generates a function name for reading a named property of a certain struct type

struct_new_call
Generates a function name for creating a certain struct type

var_name
Generate a safe variable name for a user defined function