Module fe_yulgen::runtime::functions::revert

Functions

all
Return all revert runtime functions

generate_revert_fn
Generate a YUL function to revert with data

generate_revert_fn_for_assert
Generate a YUL function to revert with the Error signature and the given set of params. NOTE: This is currently used for assert False, "message" statements which are encoded as Error(msg="message"). This will be removed in the future.

generate_revert_fn_for_panic
Generate a YUL function to revert with panic codes

generate_struct_revert
Generate a YUL function to revert with a specific struct used as error data